+++ python2verilog/api/from_context.py
@@ -7,7 +7,7 @@ from python2verilog import ir
 from python2verilog.backend import verilog
 from python2verilog.frontend.generator2ir import Generator2Graph
 from python2verilog.optimizer.optimizer import OptimizeGraph
-from python2verilog.utils.smart_asserts import get_typed
+from python2verilog.utils.assertions import get_typed
 
 
 def context_to_codegen(context: ir.Context):
