+++ python2verilog/api/from_text.py
@@ -14,7 +14,7 @@ from python2verilog.api.namespace import get_namespace
 from python2verilog.backend import verilog
 from python2verilog.frontend.generator2ir import Generator2Graph
 from python2verilog.optimizer.optimizer import OptimizeGraph
-from python2verilog.utils.smart_asserts import get_typed, get_typed_list
+from python2verilog.utils.assertions import get_typed, get_typed_list
 
 
 def text_to_verilog(
