+++ python2verilog/ir/graph.py
@@ -11,7 +11,7 @@ from __future__ import annotations
 
 from typing import Optional
 
-from ..utils.smart_asserts import get_typed, get_typed_list
+from ..utils.assertions import get_typed, get_typed_list
 from . import expressions as expr
