+++ python2verilog/utils/assertions.py
